feat: token bucket rate limiting
Replace the existing rate limiting mechanism with a token bucket
algorithm to improve request handling efficiency and control.

The previous implementation used a simple last request and current delay
calculation, which could lead to uneven request distribution when
multiple requests were made in quick succession. The result of this was
the more links discovered by as single request the more likely it was
that we would trigger 429 and 503 response codes, which combined with
max_retries would cause the crawler to fail to successfully process all
pages.

The new implementation uses a token bucket algorithm, which allows for
more flexible and efficient rate limiting. The token bucket algorithm
allows for bursts of requests while maintaining an average rate over
time. This combined with active rate limit header processing results in
a much higher success rate for crawls.

In addition the option to disable max retry limit, the new default, has
been tested to confirm 100% success rate for crawls on a site which
performs active rate limiting and previously resulted in a large number
of failures.

Finally the rate limiter starts with default rate which will be adjusted
per domain based on the rate limit headers returned by the server.